requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
markdown==3.6